Crate axruntime

Runtime library of ArceOS.

Any application uses ArceOS should link this library. It does some initialization work before entering the application’s main function.

Cargo Features

alloc: Enable global memory allocator.

paging: Enable page table manipulation support.

irq: Enable interrupt handling support.

multitask: Enable multi-threading support.

smp: Enable SMP (symmetric multiprocessing) support.

fs: Enable filesystem support.

net: Enable networking support.

display: Enable graphics support.

All the features are optional and disabled by default.

Functions

rust_main

The main entry point of the ArceOS runtime.

rust_main_secondarysmp

The main entry point of the ArceOS runtime for secondary CPUs.